+++ legate/driver/args.py
@@ -18,8 +18,8 @@ from __future__ import annotations
 
 from argparse import ArgumentDefaultsHelpFormatter, ArgumentParser
 
+from ..utils.types import LauncherType
 from . import defaults
-from .types import LauncherType
 
 __all__ = ("parser",)
