+++ legate/driver/config.py
@@ -23,8 +23,8 @@ from functools import cached_property
 from pathlib import Path
 from typing import Any
 
+from ..utils.types import ArgList, DataclassMixin, LauncherType
 from .args import parser
-from .types import ArgList, DataclassMixin, LauncherType
 from .ui import warn
 from .util import object_to_dataclass
