+++ legate/driver/driver.py
@@ -16,12 +16,12 @@ from __future__ import annotations
 
 from subprocess import run
 
+from ..utils.types import Command, EnvDict
 from .command import CMD_PARTS
 from .config import Config
 from .launcher import Launcher
 from .logs import process_logs
 from .system import System
-from .types import Command, EnvDict
 from .ui import warn
 from .util import print_verbose
