+++ legate/tester/config.py
@@ -21,9 +21,9 @@ import os
 from argparse import Namespace
 from pathlib import Path
 
+from ..utils.types import ArgList, EnvDict
 from . import DEFAULT_PROCESS_ENV, FEATURES, SKIPPED_EXAMPLES, FeatureType
 from .args import parser
-from .types import ArgList, EnvDict
 
 
 class Config:
