+++ legate/tester/stages/_linux/cpu.py
@@ -15,11 +15,8 @@
 from __future__ import annotations
 
 from itertools import chain
+from typing import TYPE_CHECKING
 
-from ... import FeatureType
-from ...config import Config
-from ...system import System
-from ...types import ArgList, EnvDict
 from ..test_stage import TestStage
 from ..util import (
     CUNUMERIC_TEST_ARG,
@@ -28,6 +25,12 @@ from ..util import (
     StageSpec,
     adjust_workers,
 )
+
+if TYPE_CHECKING:
+    from ....utils.types import ArgList, EnvDict
+    from ... import FeatureType
+    from ...config import Config
+    from ...system import System
 
 
 class CPU(TestStage):
