+++ legate/tester/stages/_linux/eager.py
@@ -14,12 +14,16 @@
 #
 from __future__ import annotations
 
-from ... import FeatureType
-from ...config import Config
-from ...system import System
-from ...types import ArgList, EnvDict
+from typing import TYPE_CHECKING
+
 from ..test_stage import TestStage
 from ..util import Shard, StageSpec, adjust_workers
+
+if TYPE_CHECKING:
+    from ....utils.types import ArgList, EnvDict
+    from ... import FeatureType
+    from ...config import Config
+    from ...system import System
 
 
 class Eager(TestStage):
