+++ legate/tester/stages/test_stage.py
@@ -21,10 +21,10 @@ from pathlib import Path
 from typing_extensions import Protocol
 
 from ...utils.colors import yellow
+from ...utils.types import ArgList, EnvDict
 from .. import PER_FILE_ARGS, FeatureType
 from ..config import Config
 from ..system import ProcessResult, System
-from ..types import ArgList, EnvDict
 from ..ui import banner, summary
 from .util import Shard, StageResult, StageSpec, log_proc
