+++ tests/unit/legate/driver/test_config.py
@@ -23,8 +23,8 @@ from pytest_mock import MockerFixture
 
 import legate.driver.config as m
 import legate.driver.defaults as defaults
-from legate.driver.types import DataclassMixin
 from legate.utils.colors import scrub
+from legate.utils.types import DataclassMixin
 
 from ...util import Capsys, powerset, powerset_nonempty
